% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate.dt.R
\name{Aggregate.dt}
\alias{Aggregate.dt}
\title{Aggregate attribute values throughout the entire tree recursively}
\usage{
Aggregate.dt(
  tree,
  ...,
  aggfun,
  traversal = c("pre-order", "post-order", "in-order", "level", "ancestor")
)
}
\arguments{
\item{tree}{the data tree object}

\item{...}{the attribute variables to be aggregated for each node}

\item{aggfun}{the aggregation function to be applied to each node's attributes}

\item{traversal}{traversal order:
any of 'pre-order' (the default), 'post-order', 'in-order', 'level', 'ancestor', or a custom function}
}
\description{
This function aggregates the selected attribute values of each node to its corresponding parent node using customised function, upto the root node.
}
\examples{
\dontrun{
# create a data tree object
data(employee)
employee.tree = createTree(employee, "Employee Information",
                           c("Department", "JobRole", "EmployeeNumber"))

# aggregate the mean of monthly income for each node recursively
# to its corresponding parent node in post-order, upto the root node
Aggregate.dt(employee.tree, "MonthlyIncome",
             aggfun = function(x) mean(x, na.rm = TRUE),
             traversal = "post-order")
}


}
