% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branches.R
\name{branches}
\alias{branches}
\title{Function to draw branch for each node}
\usage{
branches(index, selected = F, ...)
}
\arguments{
\item{index}{tree branches (lines) index}

\item{selected}{whether the vertical branch is connecting to a selected node}

\item{...}{other arguments passed on to the function}
}
\description{
Function to draw branch for each node
}
