% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branches.R
\name{branches.shifted}
\alias{branches.shifted}
\title{Function to draw branch for each node --- slight shifted version}
\usage{
branches.shifted(x0 = 0, x1 = 1, index, selected = F, ...)
}
\arguments{
\item{x0}{x0 coordinate}

\item{x1}{x1 coordinate}

\item{index}{tree branches (lines) index}

\item{selected}{whether the vertical branch is connecting to a selected node}

\item{...}{other arguments passed on to the function}
}
\description{
Function to draw branch for each node --- slight shifted version
}
