% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.split.R
\name{categorical.split}
\alias{categorical.split}
\title{Expand all factors into multiple dummy variables whilst keeping all other variables in the df}
\usage{
categorical.split(dataframe)
}
\arguments{
\item{dataframe}{a data frame containing categorical variables for expansion}
}
\value{
A data frame
}
\description{
Expand all factors (categorical variables) into multiple dummy variables whilst keeping all other variables in the df
(if the factor only contains two levels then the factor will be turned into one dummy variable with only 1/0 instead of unnecessary expanding.)
}
\examples{
### An example dataset
customers <- data.frame(
  id=c(10,20,30,40,50,60,70),
  gender=c('male','female','female','male','female', 'female', 'male'),
  mood=c('happy','sad','happy','sad','happy', "neutral", "neutral"),
  outcome=c(1,1,0,0,0,0,1))

categorical.split(customers)



}
