% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connection}
\alias{connection}
\title{Horizontal branches connecting up two levels}
\usage{
connection(index, selected = F, ...)
}
\arguments{
\item{index}{index of the middle connection branches}

\item{selected}{whether the connection branch is connecting to a selected node}

\item{...}{other arguments passed on to the function}
}
\description{
Horizontal branches connecting up two levels
}
