% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTree.R
\name{createTree}
\alias{createTree}
\title{Create data tree object from a long format data frame}
\usage{
createTree(df, root_name, hierarchy_order)
}
\arguments{
\item{df}{data frame in the long data format}

\item{root_name}{a character specifing the root name for the whole hierarchy tree}

\item{hierarchy_order}{a vector of hierarchy level names in ascending order}
}
\value{
An active-binding data tree object
}
\description{
Create data tree object from a long format data frame
}
\examples{
\dontrun{
# load the employee data
data(employee)

# create a data tree object from the employee dataset with a hierarchy order of:
# Employee Information -> Department -> JobRole -> EmployeeNumber
employee.tree = createTree(employee, "Employee Information",
           c("Department", "JobRole", "EmployeeNumber"))
}

}
