% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.dt.R
\name{display.dt}
\alias{display.dt}
\title{Print out the tree with selected attribute variables up to a certain level.}
\usage{
display.dt(tree, ..., levels.shown = tree[["height"]])
}
\arguments{
\item{tree}{the data tree object to be printed}

\item{...}{attribute variables or arguments to be passed on}

\item{levels.shown}{a numeric value indicating which hierarchy level of the tree to print up to}
}
\description{
Print out the tree with selected attribute variables up to a certain level.
}
\examples{
\dontrun{
data(employee)
employee.tree = createTree(employee, "Employee Information",
                           c("Department", "JobRole", "EmployeeNumber"))

# display the employee.tree with selected attribute variables up to the third level
display.dt(employee.tree, 3, "Age", "MonthlyIncome", "Attrition")
}


}
