% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employee.R
\docType{data}
\name{employee}
\alias{employee}
\title{Sample dataset containing employee information}
\format{
A data frame with 1470 rows and 35 variables
}
\source{
\url{https://www.kaggle.com/pavansubhasht/ibm-hr-analytics-attrition-dataset}
}
\usage{
data(employee)
}
\description{
This is a fictional dataset created by IBM data scientists.
Each row of the dataset represents an employee's information.
}
\examples{
data(employee)
}
\keyword{datasets}
