% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.plot.R
\name{group.plot_all}
\alias{group.plot_all}
\title{Hierarchical data visualisation (space-saving version)
to print the all 3 levels of nodes}
\usage{
group.plot_all(
  tree,
  x.name,
  ...,
  y.name = NULL,
  index = 1,
  levels.shown = 3,
  aspect = NULL,
  color = NULL
)
}
\arguments{
\item{tree}{a data tree object}

\item{x.name}{a character value indicating x variable}

\item{...}{ggplot functions and layers to be passed on}

\item{y.name}{a character value indicating y variable}

\item{index}{labels size and tree branches (lines) thickness}

\item{levels.shown}{a numeric value indicating which level plot up to}

\item{aspect}{plot aspect ratio}

\item{color}{a character value indicating which variable to use for colouring}
}
\value{
a svg file will be saved in the current working directory and automatically opened in a web broswer
}
\description{
Hierarchical data visualisation (space-saving version)
to print the all 3 levels of nodes
}
\examples{
\dontrun{
# load the employee data
data(employee)

# create a data tree object from the employee dataset with a hierarchy order of:
# Employee Information -> Department -> JobRole -> EmployeeNumber
employee.tree = createTree(employee, "Employee Information",
           c("Department", "JobRole", "EmployeeNumber"))


# x and y variables
group.plot_all(employee.tree, x.name ="Age", y.name ="MonthlyIncome",
               color = "EducationField", index = 2,
               levels.shown = 3, aspect = 3/4, geom_point(), geom_smooth())

# only singe variable
group.plot_all(employee.tree, x.name ="MonthlyIncome",
               color = "Gender", index = 2,
               levels.shown = 1, geom_density())

}


}
