% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\title{Draw the node names on each branch}
\usage{
labels(label, current.level = 1, no.nodes = 1, index, selected = F, ...)
}
\arguments{
\item{label}{label to be drawn as the name of node}

\item{current.level}{current hierarchy level we are drawing at}

\item{no.nodes}{number of nodes in the current level}

\item{index}{tree branches (lines) index}

\item{selected}{whether the vertical branch is connecting to a selected node}

\item{...}{other arguments passed on to the function}
}
\description{
Draw the node names on each branch
}
