% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_prep.R
\name{nested.dlply}
\alias{nested.dlply}
\title{Split a dataframe and execute a function on each splited group}
\usage{
nested.dlply(df, by, fun, ...)
}
\arguments{
\item{df}{a dataframe}

\item{by}{variables to split data frame by}

\item{fun}{function to be executed on each splited group}

\item{...}{other arguments passed on to the function}
}
\description{
Split a dataframe and execute a function on each splited group
}
