% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_prep.R
\name{plot_prep}
\alias{plot_prep}
\title{Generate ggplot objects for a certain level in the tree}
\usage{
plot_prep(df, level, x.name, y.name, color, index, aspect, ...)
}
\arguments{
\item{df}{a dataframe}

\item{level}{which level to be plotted}

\item{x.name}{x variable}

\item{y.name}{y variable}

\item{color}{variable to be coloured by}

\item{index}{labels and tree branches size index}

\item{aspect}{plot aspect ratio}

\item{...}{ggplot functions and layers to be passed on}
}
\description{
Generate ggplot objects for a certain level in the tree
}
