% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacing.plot.R
\name{spacing.plot_selected}
\alias{spacing.plot_selected}
\title{Hierarchical data visualisation (spacing version)
to print the first 2 levels + the 3rd level of selected node expanded out}
\usage{
spacing.plot_selected(
  tree,
  x.name,
  node.shown,
  ...,
  y.name = NULL,
  index = 1,
  aspect = 1,
  color = NULL
)
}
\arguments{
\item{tree}{a data tree object}

\item{x.name}{a character value indicating x variable}

\item{node.shown}{node selected in the 2nd hierarchy level to be expanded out}

\item{...}{ggplot functions and layers to be passed on}

\item{y.name}{a character value indicating y variable}

\item{index}{labels and tree branches size index}

\item{aspect}{plot aspect ratio}

\item{color}{a character value indicating which variable to use for colouring}
}
\value{
a svg file will be saved in the current working directory and automatically opened in a web broswer
}
\description{
Hierarchical data visualisation (spacing version)
to print the first 2 levels + the 3rd level of selected node expanded out
}
\examples{
\dontrun{
# load the employee data
data(employee)

# create a data tree object from the employee dataset with a hierarchy order of:
# Employee Information -> Department -> JobRole -> EmployeeNumber
employee.tree = createTree(employee, "Employee Information",
           c("Department", "JobRole", "EmployeeNumber"))


# x and y variables
spacing.plot_selected(employee.tree, x.name ="Age", y.name ="MonthlyIncome", index = 2,
                    node.shown = "Sales", geom_point(), geom_smooth())
# only singe variable
spacing.plot_selected(employee.tree, x.name ="MonthlyIncome", index = 2,
                      node.shown = "Human Resources", geom_density(),
                      ggtitle("Monthly Income density plot"))
}

}
