% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_to_df.R
\name{tree_to_df}
\alias{tree_to_df}
\title{Convert a data tree object to a long format data frame}
\usage{
tree_to_df(tree, ...)
}
\arguments{
\item{tree}{the data tree object}

\item{...}{the attributes to be added as columns in the data frame}
}
\value{
A long format data frame
}
\description{
Convert a data tree object to a long format data frame
}
\examples{
\dontrun{
# Convert the employee.tree to a long format data frame with Age and Education attribute variables
tree_to_df(employee.tree, "Age", "Education")
}

}
