% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{vertsticks}
\alias{vertsticks}
\title{Vertical branches connecting up two levels in the middle}
\usage{
vertsticks(y, segement.index, index, selected = F, ...)
}
\arguments{
\item{y}{y coordinates for each node in a vector}

\item{segement.index}{indices in y of where to draw the vertical tree branches}

\item{index}{tree branches (lines) index}

\item{selected}{whether the vertical branch is connecting to a selected node}

\item{...}{other arguments passed on to the function}
}
\description{
Vertical branches connecting up two levels in the middle
}
