% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branches.R
\name{branches}
\alias{branches}
\title{Horizontal branches connecting up two levels}
\usage{
branches(index, selected = F, ...)
}
\arguments{
\item{index}{tree branches (lines) thickness}

\item{selected}{whether the tree branch is connecting to a selected node}

\item{...}{other arguments passed on to the function}
}
\description{
Horizontal branches connecting up two levels
}
