% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branches.R
\name{branches.shifted}
\alias{branches.shifted}
\title{Horizontal branches connecting up two levels --- slight shifted version}
\usage{
branches.shifted(x0 = 0, x1 = 1, index, ...)
}
\arguments{
\item{x0}{x coordinate of where the tree branch starts}

\item{x1}{x coordinate of where the tree branch ends}

\item{index}{tree branches (lines) thickness}

\item{...}{other arguments passed on to the function}
}
\description{
Horizontal branches connecting up two levels --- slight shifted version
}
