% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_prep.R
\name{plots_prep}
\alias{plots_prep}
\title{A nested list of ggplot objects}
\usage{
plots_prep(tree, x.name, y.name, color, levels.shown, index, aspect, ...)
}
\arguments{
\item{tree}{a data tree object}

\item{x.name}{x variable}

\item{y.name}{y variable}

\item{color}{variable to be coloured by}

\item{levels.shown}{which level to be shown}

\item{index}{labels size and tree branches (lines) thickness}

\item{aspect}{plot aspect ratio}

\item{...}{ggplot functions and layers to be passed on}
}
\value{
a nested list of ggplot objects in a tree-structure
}
\description{
A nested list of ggplot objects
}
