\name{treevis-api}
\alias{treevis-api}
\alias{treevis_layout_mm}
\alias{add_gene_structure_column}
\alias{abbreviate_clade_labels}
\alias{add_alignment_column}
\alias{add_amino_acid_site_column}
\alias{add_categorical_column}
\alias{add_cluster_membership_column}
\alias{add_complete_overlap_groups}
\alias{add_fimo_column}
\alias{add_gene_cluster_membership}
\alias{add_heatmap_column}
\alias{add_integer_column}
\alias{add_long_branch_compression_marks}
\alias{add_meme_column}
\alias{add_motif_colors}
\alias{add_motif_counts}
\alias{add_node_labels}
\alias{add_node_points}
\alias{add_ortholog_column}
\alias{add_pie_charts}
\alias{add_pointplot_column}
\alias{add_protein_domain_column}
\alias{add_scale_bar}
\alias{add_signal_peptide_column}
\alias{add_peroxisome_column}
\alias{add_localization_column}
\alias{add_site_state_column}
\alias{add_symbol_site_column}
\alias{add_synteny_column}
\alias{add_synteny_similarity_column}
\alias{add_sequence_similarity_column}
\alias{add_cis_similarity_column}
\alias{add_text_column}
\alias{add_tiplabel_column}
\alias{add_tree_column}
\alias{append_branch_stats}
\alias{append_branch_tiplab_colors}
\alias{attach_neighbor_stats}
\alias{compress_long_branches_for_display}
\alias{correct_self_hit_query}
\alias{enhance_branch_table}
\alias{extract_alignment_sites}
\alias{extract_xml_attr}
\alias{geom_divtime}
\alias{get_categorical_levels}
\alias{get_df_N}
\alias{get_df_clade}
\alias{get_df_domain}
\alias{get_df_intron}
\alias{get_df_polygon}
\alias{get_df_protein_backbone}
\alias{get_df_tip}
\alias{get_df_trait}
\alias{get_domain_fill_colors}
\alias{get_label_num}
\alias{get_line_coordinate}
\alias{get_long_branch_display_settings}
\alias{get_rel_widths}
\alias{get_trait_colors}
\alias{gg_color_hue}
\alias{map_trimmed_to_untrimmed_nongap_index}
\alias{merge_overlap_group}
\alias{merge_tip_trait}
\alias{normalize_category_text}
\alias{overlay_branch_combination}
\alias{overlay_convergence}
\alias{parse_optional_numeric}
\alias{prepare_df_rps_for_plot}
\alias{propagate_tiplab_colors_to_internal_branches}
\alias{read_meme_motifs}
\alias{read_site_state_alignment}
\alias{reshape_long_base}
\alias{split_nested_clade}
\alias{tidy_df_tip}
\alias{treevis_axis_label}
\alias{treevis_extract_species_label}
\alias{treevis_font_size_factor}
\alias{treevis_ggmsa_letter_palette}
\alias{treevis_is_hybrid_binomial_connector}
\alias{treevis_is_hybrid_connector_token}
\alias{treevis_label_matches_ortholog_prefix}
\alias{treevis_ortholog_axis_label}
\alias{treevis_ortholog_prefix_target}
\alias{treevis_should_wrap_site_axis_label}
\alias{treevis_site_panel_width}
\alias{treevis_site_state_palette}
\alias{treevis_species_prefix_token_count}
\alias{treevis_strip_ortholog_prefix}
\alias{treevis_taxonomic_token_key}
\alias{treevis_text_size}
\alias{treevis_tidy_msa}
\alias{treevis_wrap_axis_label}
\alias{truncate_panel_text}
\title{GeneGalleon tree visualization API}
\description{
Internal plotting, annotation, alignment, motif, and tree-layout helpers used by
GeneGalleon's R workflow entry points. These functions are exported so the
entry-point scripts can use a stable installed-package boundary.
}
\details{
The API is intended for GeneGalleon workflow code. Inputs are generally the
plot state and parsed workflow arguments described by the corresponding
entry-point script.
}
\keyword{internal}
